v3.0.0 Date of change:Saturday, September 17, 2022
Changes:
- Significantly changed the design
- Eliminated Save to XML and Load from XML
- Added save to JSON and load from JSON functionality

Notes:
The XML saved by the previous "Save from XML" function will not be usable in future versions. However, you can solve this problem by converting to JSON, for example by searching the Internet for "online XML to JSON conversion". Please note that we are not responsible for any problems that may occur when converting from XML to JSON online.